pyproject.toml
[project]
name = "sales-coach-demo"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "anthropic>=0.40.0",
    "websockets>=13.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]
